+++ MangAdventure/tests/test_middleware.py
@@ -1,10 +1,7 @@
 from random import randint
 
-from django.conf import settings
 from django.test import Client
 from django.urls import reverse
-
-from pytest import mark
 
 from MangAdventure.bad_bots import BOTS
 
@@ -26,15 +23,3 @@ class TestBaseMiddleware(MangadvTestBase):
     def test_early_data(self):
         r = self.client.get(reverse('index'), HTTP_EARLY_DATA='1')
         assert r.status_code == 425
-
-
-@mark.skipif(
-    settings.ACCOUNT_DEFAULT_HTTP_PROTOCOL != 'https',
-    reason='PreloadMiddleware requires HTTPS'
-)
-class TestPreloadMiddleware(MangadvTestBase):
-    def test_preload(self):
-        r = self.client.get(reverse('index'))
-        assert 'style.css>; as=style' in r['Link']
-        r = self.client.get(reverse('opensearch'))
-        assert 'Link' not in r