+++ docs/changelog.rst
@@ -6,6 +6,7 @@ v0.9.1
 
 * Fixed stale cache on new chapter
 * Added XML metadata to chapter archives
+* Removed HTTP/2 push middleware
 
 v0.9.0
 ^^^^^^